refactor: replace domain-specific models with standard types and improve test consistency
- Replaced `UserId` and `Username` with `long` and `String` respectively across codebase.
- Updated session and token processing logic in `JwtAuthenticationFilter` for better handling.
- Refactored existing tests (`ApplicationConfigTest`, `UserApplicationServiceTest`, `AuthControllerTest`) to align with new model changes and improve readability.
- Simplified mocked object creation and dependency injection in tests.

+++ src/test/java/io/vitormmartins/chatforge/ChatForgeApplicationTests.java
@@ -67,7 +67,7 @@ class ChatForgeApplicationTests {
     private final String testUsername = "Capivara";
     private final String testPassword = "aravipac";
     private final String testEmail = "capivara@example.com";
-    private final Integer testUserId = 11;
+    private final Long testUserId = 11L;
     private final LocalDateTime testCreatedAt = LocalDateTime.now();
 
     @BeforeEach
